refactor: split utility code
