+++ src/utils/movie.js
@@ -1,19 +1,5 @@
-import {ONE_HOUR} from "./const";
-
-export const getRandomFractionalNumber = (a = 1, b = 0) => {
-  const lower = Math.min(a, b);
-  const upper = Math.max(a, b);
-  return lower + Math.random() * (upper - lower);
-};
-
-export const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-export const getRandomArbitrary = (min, max) => Math.random() * (max - min) + min;
+import {ONE_HOUR} from "../const";
+import {getRandomInteger} from "./common";
 
 export const getRandomElement = (array) => array[getRandomInteger(0, array.length - 1)];
