+++ src/view/movie-list-container.js
@@ -1,1 +1,1 @@
-export const createMoviesListContainerTemplate = ()=> `<div class="films-list__container"></div>`;
+export const createMoviesListContainerTemplate = () => `<div class="films-list__container"></div>`;
